refactor(cli): extract manifest runtime config domain
